[#151572737] Test Environment Setup in Yojimbo
Setting up the Skillz version test environment for Game Sync. Changed all the unit test, server, and client files to use the Skillz messages instead. Modified the object type for message to Skillz messages in files that use the original test messages. Modified the `premake5.lua` to make build project with our messages header file `message.h`. Run all the projects with the new setup with premake5 commands and they all are successfully built. Run cppcheck and scan-build to make sure no extra issues occur. Run the unit tests and they all pass with the Skillz messages. Run soak and messages are successfully send from the server to client and from client to server. Run the server and client and see they are successfully start up and the client is able to connect to the server without problem. Multiple clients are able to connect to the the server at once successfully. The valgrind shows no leaks. All the premake5 commands can be run successfully with the new changes.